Fix(evaluator): Extract built-in functions into builtInFuns.cpp
- It can make the code clearer
